Implement a large refactor of the developer's API.
In order to prevent DRY violations by rewriting model validation in
javascript and in python, editor related templates will be rendered
serverside.  It is still encouraged to return {{ mustache }} templates,
but widgets should supply an ``edit_template`` and a
``preview_template`` for use with the Backbone widgy application.

Sane defaults are defined and there are a few options for where to place
your templates (See ``widgy.models.get_preview_template`` and
``widgy.models.edit_templates``)

When form validation errors occur, clients will recieve a 409 CONFLICT
and a JSON object of field -> messages mapping.

As of now, error validation messages are rendered clientside, but we may
want to look into doing this serverside as well and return HTML
replacements instead of JSON.

-   Add a `get_form_class` method on Content models.  Defaults to a
    basic ModelForm.

    Since `get_form_class` is a wrapper around Django's
    `modelform_factory` we need to look into memoizing this method.  It
    **may** be a concern that this creates a new class as views that
    will access this method will be requested many times.

-   Generate edit and preview templates serverside and return them to
    the client as data, part of the JSON response when retrieving
    Content models.

-   Add a CMSContent abstract model class for CMS content models to
    inherit.  This allows overwriting of the ``get_render_templates``
    method.