implement WidgetConfig as a Proxy.
When you write a WidgetConfig, it will be mixed into the model class and
when dealing with validation, the WidgySite will use the proxy
WidgetConfig.  By default, WidgetConfig implements nothing except a
delegation method.  But you can incrementally implement methods that you
need.  For example, the following implementation of bucket will accept
anything.

    from widgy.models import Content
    from widgy import registry

    class Bucket(Content):
        pass

    registry.register(Bucket)

But if you wanted it to be more selective, you could subclass it:

    class VowelBucket(Bucket):
        """
        VowelBucket only accepts classes whose name starts with vowels.
        """
        def valid_parent_of(self, class, obj=None):
            return class.__name__[0].lower() in 'aeiou'

    registry.register(VowelBucket)

However, subclassing it like this means that you would have two tables
in the database, and really none of the fields have changed.  Instead,
you could use a WidgetConfig to override the compatibility checking.

    from widgy import WidgetConfig, registry

    class VowelChildConfig(WidgetConfig):
        def valid_parent_of(self, class, obj=None):
            return class.__name__[0].lower() in 'aeiou'

    registry.register(Bucket, VowelChildConfig)

This has the added benefit of being reusable.  For example, you could
make all of your Content classes only accept children whose classes
start with a vowel:

    registry.register(OtherContainer, VowelChildConfig)