Bugfix/optimization for shelf compatibility after tree modifications.
Provides an optimization for refreshing the shelf compatibility.
Previously, when editing a node, you had to do two requests (one for
the node, one for the shelf compatibility) to update the UI.  In
addition, the shelf request had to happen after the node one, to
prevent getting the compatibility wrong.  This refreshes compatibility
in one request instead of waiting for two round trips.