Add pre_delete
Content nows has a `pre_delete` method, that is called before a widget
is deleted but while it is still in the tree. Also moves the canonical
`delete` method onto Content, with a new `raw` argument. `raw=False` is
used to delete a widget without calling `pre_delete`, which is required if
the widget was created without calling `post_create`. This happens when
adding a widget that turns out to be invalid.