Node.frozen -- prohibits all mutations to a widget
When a Node is frozen, any modification to it or its Content is
prohibited. This is used to prevent old versions from being edited in
any way.

Also, VersionTracker.commit copies the working copy into the commit,
instead of using the working copy for the commit and making a new
working copy. This prevents the frontend from having to refresh its tree
every commit, instead it can just keep using the same one.