Uncaptcha widget
Adds a get_form_mixins method fields. It allows a field to add a mixin
class to the inheritance hierarchy of the form its in.

I split FormField into two classes - BaseFormField and FormField.
Uncaptcha is a BaseFormField but not a FormField, because it doesn't
have a value. I'm not sure how exactly to express this.