Remove the mezzanine_urls.py file.
We previously used this to enable you to use the mezzanine.blog app, but
in conjunction with urlconf_include.  This created a problem because
mezzanine updates the file that this is based on, which causes
difficulties when upgrading your mezzanine version.

We no longer support installing mezzanine.blog with urlconf_include,
because we rely on the standard mezzanine urls file.