WidgyField.render
The entry-point for rendering a Widgy tree is now a method on the
WidgyField. This allows use to add the `widgy` template context variable
to hold metadata about the Widgy field. It is a dictionary containing:

  - `site` -- The Widgy site for the field.
  - `owner` -- The model instance whose Widgy field we are currently
     rendering.
  - `parent` -- The parent `widgy` object when rendering a nested Widgy
     field.

The current motivation for this change is to allow Widgy forms to ask
their owner which URL to submit to. Having the Widgy site available is
not used currently, but is expected to be useful.

The presence of `request` in the context is no longer checked during
rendering.