rm the requirement that widgy models can be imported before there's a widgy site
We used to require that widgy models could be imported at the top of the widgy
site file, before the site existed. We're removing this requirement to allow
fields to access the site. Since the widgy site _is_ widgy, widgy doesn't
really exist before there's a site. This means it doesn't make sense to access
widgy models before widgy itself exists.

This means that fields can use the site -- we'll use this for
VersionedWidgyField to get the version tracker model.