Changed EmailUserHandler.to to reference an ident
EmailSuccessHandler has a foreign key to a FormInput; however, the foreign
key is never updated when a tree is committed, so it fails when executing.
This fixes that problem by not doing having FKs to nodes in this instance.
We also maintained the on_delete=PROTECT behavior by using the new
pre_delete_widget signal.