Automatically include admin scss files
Each installed widget gets a chance to include an admin scss for itself
or its app. These scss files will be included when rendering the widgy
editor.

The path for an app looks like

    widgy/page_builder/admin.scss

and for a widget:

    widgy/page_builder/table.admin.scss

This will help us fix the issue where css for individual projects
gets committed in widgy core. This is not entirely backwards
compatible--icons were removed for widgets not provided by widgy.

TODO: there is still `page_builder.Table` css in
`widgy/css/widgy.scss`. This will need to be removed and put in
`widgy/page_builder/table.admin.scss`.