Add 'owner' links on the history page and the form submissions page
Add a link to the admin change page of each owner. This can not be
merged as-is --- the form admin hardcodes a reference to
ReviewedVersionTracker. It needs access to a site so it can call
get_version_tracker_model.

(cherry picked from commit a884da03b319b0f45dd6bf7dca9e579d219c6524)

Merge #86

Reviewed-By: Antoine Catton <acatton@fusionbox.com>