Fix the memory leak in urlconf_include
Django's urlresolvers.get_resolver function is memoized, causing a
memory leak when a new urlconf module is used for every request.

The other way to do this would be to use the same urlconf module for
every request. This would require caching the instances based on the
current set of UrlconfIncludePages.