User delete querying nonexistent reviewqueue
A widgy project was searching for user-related review queue objects when an
admin attempted to delete said user. Thus, the site was attempting to query a
database tab le that did not exist.

Reorganizing the review queue imports within the widgy-mezzanine admin file so
that they are not called unless review queue is in INSTALLED_APPS fixed the
problem