Fix inheritance for LinkFields
Don't copy the link_registry when copying a LinkField. When inheriting
from a model with a LinkField, each field from the parent is deepcopied
and attached to the child. This means that the child gets a different
instance of its link_registry, and could not learn about models
registered later.