Restructured the backend SCSS.
I have split up widgy.scss into two files, widgy_layout.scss and
widgy_theme_default.scss in the hope that it will be easier to edit the
styles of the editor without affecting the functionality and to reduce
some specificity headaches that we were having.  I have also included a
super barebones theme called widgy_theme_lofi.scss for testing the
structural SCSS.

I'm sure there are regressions in the CSS, I was hoping to get a real
designer to fix those because they are beyond my power.