Fix ReviewedVersionTrackerQuerySet.published
All of the published conditions must be in the same call to filter, so
we can't call super(). Otherwise if the tracker has a commit that is
published but not approved, and another commit that is approved but not
published, the tracker will be counted as published even though no one
of its commits is.

See
<https://docs.djangoproject.com/en/dev/topics/db/queries/#spanning-multi-valued-relationships>
for the difference.

The other option would be to try to use the undocumented
_next_is_sticky() method, but this doesn't work because
VersionTrackerQuerySet includes a call to distinct(), which resets the
sticky flag.