Fix large file upload bug
When large files are uploaded, they are stored as temp files and moved
when the storage.save method is called. This method was called in
FileUpload's serialize_value method, so if serialize_value was called
more than once (ex: by the EmailSuccessHandler), the temp file would
be moved in the first call, and would not exist for subsequent calls,
resulting in an IOError.